Move camera info to its own file
